Added additional tongue logic, minor collision
The tongue now shoots outs without leaving the screen, it also retracts without leaving the scree. The tongue has multiple states which allow it to extend and retract properly. Although, I do not like the way I implemented this using a multitude of conditional statements. I also added collision with the fly by making the fly a global object so that my function to detect the collision may make use of the fly's position to measure the distance between the tongue and the fly.