Used Pippin's Frogfrogfrog
I was satisfied with the progress I had made, but I was originally gonna re-write the code so I could have a deeper understanding of what was happening. Enough time passed that I forgot my own code. It was messy any who so I went back to using Pippin's Frogfrogfrog.