Additional soundeffects
I added two more sound effects, changed one, and added a filter to the music during the game over state. A sound plays when the player loses, simulating Hanez falling in the water, another for when the player plays again and I changed the sound when the tongue hits the thorns. A muffled the music during the game over screen to give the effect that Hanez is underwater. I also deleted a few unecesary assets from the library.