Creating Celestial Bodies
I am now implementing the general structure of the celestial bodies. I am going to be using vectors to simulate the motion and gravitational pull the sun has on the celestials. I am also adding an array so that I can later add a trail to the celestial bodies simply because this will act to the fun of watching these bodies orbit around the sun.