Draw sun
Added the draw sun function, which will take our sun object and draw a circle which follows the mouse. I also added dotted lines because I like the way it looks. This is where I will need to change the passed object to allow the player to change the body they are controlling. I would want the dotted lines to act like net, and if the user clicks on a body, the net could surround the body and "caputre" it.