Added timeout mechanic + working on supernova
I added a mechanic that times out celestials that are off screen for too long to save on memory. I'm also trying to work on a supernova mechanic that pushes away all celestials with great force. This would not only be a very interesting mechanic but could double as a screen wipe. if all Celestials are pushed far off screen they will timeout and be removed from the celestials array.